+++ fltk-derive/src/display.rs
@@ -15,6 +15,10 @@ pub fn impl_display_trait(ast: &DeriveInput) -> TokenStream {
         format!("{}_{}", name_str, "set_buffer").as_str(),
         name.span(),
     );
+    let get_style_buffer = Ident::new(
+        format!("{}_{}", name_str, "get_style_buffer").as_str(),
+        name.span(),
+    );
     let text_font = Ident::new(
         format!("{}_{}", name_str, "text_font").as_str(),
         name.span(),
@@ -228,6 +232,18 @@ pub fn impl_display_trait(ast: &DeriveInput) -> TokenStream {
                 }
             }
 
+            fn style_buffer(&self) -> Option<TextBuffer> {
+                unsafe {
+                    assert!(!self.was_deleted());
+                    let buffer = #get_style_buffer(self._inner);
+                    if buffer.is_null() {
+                        None
+                    } else {
+                        Some(TextBuffer::from_ptr(buffer))
+                    }
+                }
+            }
+
             fn text_font(&self) -> Font {
                 assert!(!self.was_deleted());
                 assert!(self.buffer().is_some());
@@ -382,7 +398,7 @@ pub fn impl_display_trait(ast: &DeriveInput) -> TokenStream {
                 }
             }
 
-            fn set_highlight_data(&mut self, mut style_buffer: TextBuffer, entries: Vec<StyleTableEntry>) -> crate::text::StyleTables {
+            fn set_highlight_data(&mut self, mut style_buffer: TextBuffer, entries: Vec<StyleTableEntry>) {
                 assert!(!self.was_deleted());
                 assert!(self.buffer().is_some());
                 assert!(entries.len() < 29);
@@ -395,8 +411,7 @@ pub fn impl_display_trait(ast: &DeriveInput) -> TokenStream {
                     sizes.push(entry.size as i32);
                 }
                 unsafe {
-                    let x = #set_style_table_entry(self._inner, style_buffer.as_ptr() as *mut raw::c_void, &mut colors[0], &mut fonts[0], &mut sizes[0], entries.len() as i32);
-                    StyleTables { _inner: x }
+                    #set_style_table_entry(self._inner, style_buffer.as_ptr() as *mut raw::c_void, &mut colors[0], &mut fonts[0], &mut sizes[0], entries.len() as i32)
                 }
             }