+++ fltk-sys/build.rs
@@ -47,6 +47,8 @@ fn main() {
         println!("cargo:rerun-if-changed=cfltk/include/cfl.h");
         println!("cargo:rerun-if-changed=cfltk/include/cfl_widget.h");
         println!("cargo:rerun-if-changed=cfltk/include/cfl_group.h");
+        println!("cargo:rerun-if-changed=cfltk/include/cfl_input.h");
+        println!("cargo:rerun-if-changed=cfltk/include/cfl_output.h");
         println!("cargo:rerun-if-changed=cfltk/include/cfl_window.h");
         println!("cargo:rerun-if-changed=cfltk/include/cfl_button.h");
         println!("cargo:rerun-if-changed=cfltk/include/cfl_box.h");
@@ -73,6 +75,8 @@ fn main() {
         println!("cargo:rerun-if-changed=cfltk/src/cfl_misc.cpp");
         println!("cargo:rerun-if-changed=cfltk/src/cfl_text.cpp");
         println!("cargo:rerun-if-changed=cfltk/src/cfl_image.cpp");
+        println!("cargo:rerun-if-changed=cfltk/src/cfl_input.cpp");
+        println!("cargo:rerun-if-changed=cfltk/src/cfl_output.cpp");
         println!("cargo:rerun-if-changed=cfltk/src/cfl_draw.cpp");
         println!("cargo:rerun-if-changed=cfltk/src/cfl_table.cpp");
         println!("cargo:rerun-if-changed=cfltk/src/cfl_tree.cpp");
@@ -82,6 +86,12 @@ fn main() {
             .args(&["submodule", "update", "--init"])
             .current_dir(manifest_dir.clone())
             .status()
+            .expect("Git is needed to retrieve the fltk source files!");
+
+        Command::new("git")	
+            .args(&["checkout", "master"])	
+            .current_dir(manifest_dir.join("cfltk").join("fltk"))	
+            .status()	
             .expect("Git is needed to retrieve the fltk source files!");
 
         if cfg!(feature = "fltk-shared") {