+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # Changelog
 
+## [0.8.8] - 2020-09-19
+### Changes
+- Fix typos, thanks @nicolasbauw 
+- Add DisplayExt::style_buffer().
+- Fix wrapper bug which ignored return values.
+- Update FLTK.
+- Impl Send and Sync for FltkError.
+- Remove opaque type StyleTables.
+- BrowserExt::set_column_widths() takes a static slice.
+
 ## [0.8.7] - 2020-09-14
 ### Changes
 - Accept AsRef<Path> where &Path was previously required.